+++ imports/helpers/Pdf/bill/not-service/index.js
@@ -1,0 +1,5 @@
+export default function notService () {
+  return [
+    {text: 'Operação não sujeita a emissão de Nota Fiscal de Serviços, nos termos da Lei Complementar Número 116/03 de 31 de Julho de 2003 que exclui a Locação de Bens Móveis da Lista de Serviços.', style: 'p'}
+  ]
+}